Moves fallback warning to utils / adds test [Run CI]
